requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colors.py==0.2.2
cryptography==2.9.2
discord.py==1.5.1
docxcompose==1.3.2
Flask==1.1.2
google-api-core==1.22.1
google-api-python-client==2.1.0
google-auth==1.20.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.52.0
gunicorn==20.1.0
httplib2==0.18.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
multidict==4.7.5
numpy==1.20.1
oauthlib==3.1.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.3.0
pypandoc==1.5
PyPDF2==1.26.0
python-docx==0.8.11
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.14.0
soupsieve==2.2.1
uritemplate==3.0.1
urllib3==1.25.10
websockets==8.1
Werkzeug==1.0.1
yarl==1.4.2
